unify unCapitalize methods
